Dia10 Meu orientador me presenteou com os coeficientes das principais wavelets possibilitando assim a implementação do algoritmo de mallat, o qual, foi, aparentemente feito com sucesso! O orientador também mandou um exemplo de sinal transformado/original e o teste que isso proporcionou fez com que houvesse quase certeza que o código funciona! yay! Pendências Preciso ainda de mais mais bases de testes (pedir mais para o orientador, se possível) Preciso saber porquê a implementação no octave dá resultados diferentes (qual seria a diferença? Normalização?).

+++ src/lib/wavelet/waveletCoeficients.cpp
@@ -9,8 +9,8 @@
 
 namespace wavelets {
 
-	std::vector<double> calcOrthogonal(std::vector<double> &vector) {
-		std::vector<double> result(vector.size());
+	std::vector<long double> calcOrthogonal(std::vector<long double> &vector) {
+		std::vector<long double> result(vector.size());
 		double multiplier = 1;
 
 		for (int index = vector.size() - 1; index >= 0; index--) {
